Move window modules into common folder
