+++ src/show-notification/show-trx-tax-report-notification.js
@@ -3,7 +3,7 @@
 const PROCESS_MESSAGES = require(
   '../../bfx-reports-framework/workers/loc.api/process.message.manager/process.messages'
 )
-const wins = require('../windows')
+const wins = require('../window-creators/windows')
 const { isWindowInvisible } = require('../helpers/manage-window')
 const showNotification = require('./')
