+++ src/window-creators/layouts/app-init-error.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link rel="shortcut icon" href="../../bfx-report-ui/build/favicon.ico">
-    <link href="../../bfx-report-ui/build/fonts/roboto.css" rel="stylesheet">
+    <link rel="shortcut icon" href="../../../bfx-report-ui/build/favicon.ico">
+    <link href="../../../bfx-report-ui/build/fonts/roboto.css" rel="stylesheet">
     <title>Bitfinex Reports</title>
     <style>
       body {
