+++ src/window-creators/layouts/app-init.html
@@ -4,8 +4,8 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-  <link rel="shortcut icon" href="../../bfx-report-ui/build/favicon.ico">
-  <link href="../../bfx-report-ui/build/fonts/roboto.css" rel="stylesheet">
+  <link rel="shortcut icon" href="../../../bfx-report-ui/build/favicon.ico">
+  <link href="../../../bfx-report-ui/build/fonts/roboto.css" rel="stylesheet">
   <title>Bitfinex Reports</title>
   <style>
     html {
@@ -201,7 +201,7 @@
     <div></div>
   </div>
   <div class="logo">
-    <img class="logo-img" src="../../build/loader.png" alt="logo">
+    <img class="logo-img" src="../../../build/loader.png" alt="logo">
   </div>
   <div id="description" class="description"></div>
 </body>
